refactor(state): extract shell options into separate module
